chore(warehouse-sources): split delta maintenance out of DeltaTableHelper
Phase 3 (PR 1) of the warehouse-sources untangling: DeltaTableHelper carried
four jobs in one 1,200-line file. This extracts the maintenance concern and
the shared primitives into a new pipelines/core/delta/ package:

- delta/errors.py: the transient-error classifiers, moved verbatim
- delta/ops.py: execute_with_conflict_retry as a free function, shared by
  the write merges, the SCD2 close merge, and compaction
- delta/maintenance.py: DeltaMaintenance, a stateless wrapper over the
  helper holding compact/vacuum/fragmentation and the vacuum-watermark
  cadence

DeltaMaintenance.run_scheduled now owns watermark persistence: both call
sites (the pre-write defensive pass in each pipeline and the CDC post-load
pass) previously duplicated the read-watermark/run/persist block with
subtly divergent transient-error checks. The narrower
is_transient_object_store_error check on the post-load paths is unified to
is_transient_maintenance_error, so the concurrent-maintenance file race and
app-DB connection blips stop minting error-tracking issues there too.

extract.run_pre_write_defensive_compact is deleted (both pipelines call
run_scheduled directly) and the unused vacuum_table method is dropped.
The dlt import stays with the writer, so maintenance-only callers no
longer pay for it. No behavior changes beyond the error-classification
unification above.
