+++ products/warehouse_sources/backend/temporal/data_imports/pipelines/common/extract.py
@@ -19,11 +19,10 @@ from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.arr
     DuplicatePrimaryKeysException,
 )
 from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.cdp_producer import CDPProducer
-from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.delta_table_helper import (
-    DeltaTableHelper,
-    is_transient_maintenance_error,
+from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.delta.errors import (
     is_transient_object_store_error,
 )
+from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.delta_table_helper import DeltaTableHelper
 from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.person_property_row_sink import (
     PersonPropertyRowSink,
 )
@@ -697,54 +696,3 @@ async def person_property_sink_clear_chunks(sink: PersonPropertyRowSink):
 async def stage_chunk_for_person_property_sink(sink: PersonPropertyRowSink, index: int, pa_table: pa.Table):
     if await sink.should_stage():
         await sink.stage_chunk(chunk=index, table=pa_table)
-
-
-async def run_pre_write_defensive_compact(
-    delta_table_helper: DeltaTableHelper,
-    schema: "ExternalDataSchema",
-    resource: SourceResponse,
-    logger: FilteringBoundLogger,
-) -> None:
-    """Best-effort pre-write compact + vacuum at the start of a sync run.
-
-    Delegates to `DeltaTableHelper.run_maintenance`, which compacts a fragmented Delta
-    target (a sync that arrived fragmented because earlier attempts failed before
-    reaching `_post_run_operations` — keeping the subsequent per-partition merge scans
-    cheap) and otherwise vacuums on a commit-count cadence so a table that OOMs its merge
-    every run and never reaches post-load compaction still sheds tombstones (the
-    ~99%-dead-file tables). The helper returns the single vacuum watermark to persist;
-    the CDC post-load path in `common/load.py` writes the same watermark, and both merge
-    via `update_sync_type_config_keys` under a row lock. Wrapped in try/except so a
-    maintenance failure never blocks the actual sync; the original error path is unaffected.
-    A transient infra error (see `is_transient_maintenance_error`) — an object-store hiccup, a racy
-    concurrent-maintenance DeltaError, or an app-DB connection blip — is logged at warning instead of
-    captured. The next sync's maintenance pass retries it from scratch, so it isn't a bug in this
-    function, just a temporary blip talking to our own S3 bucket, delta table, or app DB.
-
-    Used by both `PipelineNonDLT.run` (v2) and `PipelineV3.run` to keep the behaviour
-    identical across pipelines without each having to know how to look up `partition_count`
-    or how to swallow maintenance errors.
-    """
-    try:
-        from products.warehouse_sources.backend.models.external_data_schema import (  # noqa: PLC0415 — Django model import kept off this activity module's load path
-            update_sync_type_config_keys,
-        )
-
-        partition_count_for_compact = schema.partition_count or resource.partition_count
-        last_vacuum_version = schema.last_vacuum_version
-        commit_threshold = settings.DATA_WAREHOUSE_VACUUM_COMMIT_THRESHOLD
-        new_version = await delta_table_helper.run_maintenance(
-            partition_count=partition_count_for_compact,
-            last_vacuum_version=last_vacuum_version,
-            commit_threshold=commit_threshold,
-        )
-        if new_version is not None and new_version != last_vacuum_version:
-            await database_sync_to_async_pool(update_sync_type_config_keys)(
-                schema.id, schema.team_id, updates={"last_vacuum_version": new_version}
-            )
-    except Exception as e:
-        if is_transient_maintenance_error(e):
-            await logger.awarning(f"Pre-write maintenance skipped: transient infra error: {e}")
-            return
-        capture_exception(e)
-        await logger.aexception(f"Pre-write maintenance failed: {e}", exc_info=e)
