+++ products/warehouse_sources/backend/temporal/data_imports/pipelines/common/test/test_extract.py
@@ -5,9 +5,6 @@ import pytest
 from posthog.test.base import BaseTest
 from unittest.mock import AsyncMock, MagicMock, patch
 
-from django.db import InterfaceError, OperationalError
-
-import deltalake.exceptions
 from asgiref.sync import async_to_sync
 from parameterized import parameterized
 
@@ -21,7 +18,6 @@ from products.warehouse_sources.backend.temporal.data_imports.pipelines.common.e
     persist_primary_keys,
     report_heartbeat_timeout,
     resolve_primary_keys,
-    run_pre_write_defensive_compact,
 )
 from products.warehouse_sources.backend.temporal.data_imports.util import NonRetryableException
 
@@ -128,138 +124,6 @@ class TestPersistPrimaryKeys:
             await persist_primary_keys(schema, resource, True, logger)
 
         logger.aexception.assert_awaited_once()
-
-
-class TestRunPreWriteDefensiveCompact:
-    @parameterized.expand(
-        [
-            # (schema_partition_count, resource_partition_count, expected_passed_to_run_maintenance)
-            ("schema_value_wins", 10, 72, 10),
-            ("falls_back_to_resource", None, 72, 72),
-            ("both_none_passes_none", None, None, None),
-        ]
-    )
-    @pytest.mark.asyncio
-    async def test_resolves_partition_count_schema_over_resource(
-        self, _name: str, schema_count: int | None, resource_count: int | None, expected: int | None
-    ):
-        run_maintenance = AsyncMock(return_value=None)
-        helper = MagicMock(run_maintenance=run_maintenance)
-
-        await run_pre_write_defensive_compact(
-            helper,
-            MagicMock(partition_count=schema_count, sync_type_config={}),
-            MagicMock(partition_count=resource_count),
-            MagicMock(aexception=AsyncMock()),
-        )
-
-        assert run_maintenance.await_args is not None
-        assert run_maintenance.await_args.kwargs["partition_count"] == expected
-
-    @pytest.mark.asyncio
-    async def test_swallows_maintenance_failure(self):
-        # The whole point of the wrapper: a maintenance error must never propagate and
-        # block the sync — it's captured and logged instead.
-        helper = MagicMock(run_maintenance=AsyncMock(side_effect=RuntimeError("maintenance blew up")))
-        logger = MagicMock(aexception=AsyncMock())
-
-        schema = MagicMock(partition_count=5, sync_type_config={})
-        with patch(f"{_EXTRACT_MODULE}.capture_exception") as mock_capture:
-            await run_pre_write_defensive_compact(helper, schema, MagicMock(partition_count=None), logger)
-
-        mock_capture.assert_called_once()
-        logger.aexception.assert_awaited_once()
-
-    @parameterized.expand(
-        [
-            (
-                "credentials_loading",
-                OSError,
-                "Operation not supported: an error occurred while loading credentials: dispatch failure: timeout",
-            ),
-            (
-                "credential_provider_not_enabled",
-                OSError,
-                "Operation not supported: the credential provider was not enabled: no providers in chain provided credentials",
-            ),
-            (
-                "generic_s3_error",
-                OSError,
-                "Generic S3 error: Error getting list response body: operation timed out",
-            ),
-            (
-                # table.vacuum()/optimize.compact() surface the identical object-store error text
-                # wrapped in DeltaError instead of OSError (unlike is_deltatable()'s OSError) — the
-                # exact shape of the issue this test guards against.
-                "generic_s3_error_as_delta_error",
-                deltalake.exceptions.DeltaError,
-                "Generic error: Kernel error: Error interacting with object store: Generic S3 error: "
-                "Server returned non-2xx status code: 503 Service Unavailable: SlowDown",
-            ),
-        ]
-    )
-    @pytest.mark.asyncio
-    async def test_logs_transient_object_store_error_without_capturing(
-        self, _name: str, error_cls: type[Exception], error_message: str
-    ):
-        # A transient blip talking to our own delta S3 bucket (credential-provider or connectivity
-        # errors from delta-rs) isn't a bug in this function — it shouldn't flood error tracking the
-        # way an actual maintenance bug does (see test_swallows_maintenance_failure above).
-        helper = MagicMock(run_maintenance=AsyncMock(side_effect=error_cls(error_message)))
-        logger = MagicMock(aexception=AsyncMock(), awarning=AsyncMock())
-
-        schema = MagicMock(partition_count=5, sync_type_config={})
-        with patch(f"{_EXTRACT_MODULE}.capture_exception") as mock_capture:
-            await run_pre_write_defensive_compact(helper, schema, MagicMock(partition_count=None), logger)
-
-        mock_capture.assert_not_called()
-        logger.awarning.assert_awaited_once()
-        logger.aexception.assert_not_awaited()
-
-    @pytest.mark.asyncio
-    async def test_logs_transient_delta_maintenance_race_without_capturing(self):
-        # Regression: a concurrent optimize/vacuum pass on the same table (e.g. a zombie Temporal
-        # attempt racing its own retry) can have `optimize.compact` scan a file the other attempt
-        # already vacuumed away. Nothing gets committed when the scan fails, so the table isn't
-        # corrupted — this must be treated the same as the object-store blips above, not captured.
-        error = deltalake.exceptions.DeltaError(
-            "Failed to parse parquet: Optimize selected-file scan failed while scanning data: "
-            "Object at location .../part-0.parquet not found: 404 Not Found"
-        )
-        helper = MagicMock(run_maintenance=AsyncMock(side_effect=error))
-        logger = MagicMock(aexception=AsyncMock(), awarning=AsyncMock())
-
-        schema = MagicMock(partition_count=5, sync_type_config={})
-        with patch(f"{_EXTRACT_MODULE}.capture_exception") as mock_capture:
-            await run_pre_write_defensive_compact(helper, schema, MagicMock(partition_count=None), logger)
-
-        mock_capture.assert_not_called()
-        logger.awarning.assert_awaited_once()
-        logger.aexception.assert_not_awaited()
-
-    @parameterized.expand(
-        [
-            ("dns_resolution_failure", OperationalError, "[Errno -2] Name or service not known"),
-            ("pooler_dropped_connection", InterfaceError, "connection already closed"),
-        ]
-    )
-    @pytest.mark.asyncio
-    async def test_logs_transient_db_connection_error_without_capturing(
-        self, _name: str, error_cls: type[Exception], error_message: str
-    ):
-        # A DNS/pooler blip hit while resolving `job.folder_path()` on a pooled app-DB connection
-        # (e.g. inside `_get_delta_table_uri`) isn't a maintenance bug either — same treatment as
-        # the object-store blips above, so a self-healing retry doesn't flood error tracking.
-        helper = MagicMock(run_maintenance=AsyncMock(side_effect=error_cls(error_message)))
-        logger = MagicMock(aexception=AsyncMock(), awarning=AsyncMock())
-
-        schema = MagicMock(partition_count=5, sync_type_config={})
-        with patch(f"{_EXTRACT_MODULE}.capture_exception") as mock_capture:
-            await run_pre_write_defensive_compact(helper, schema, MagicMock(partition_count=None), logger)
-
-        mock_capture.assert_not_called()
-        logger.awarning.assert_awaited_once()
-        logger.aexception.assert_not_awaited()
 
 
 class TestReportHeartbeatTimeoutRecording(BaseTest):
