+++ products/warehouse_sources/backend/temporal/data_imports/pipelines/common/test/test_load.py
@@ -16,6 +16,7 @@ from products.warehouse_sources.backend.temporal.data_imports.pipelines.common.l
     run_post_load_operations,
     update_job_row_count,
 )
+from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.delta.maintenance import DeltaMaintenance
 
 _LOAD_MODULE = "products.warehouse_sources.backend.temporal.data_imports.pipelines.common.load"
 _DB_RETRY_MODULE = "products.warehouse_sources.backend.temporal.data_imports.pipelines.common.db_retry"
@@ -39,12 +40,10 @@ def _make_schema(*, is_cdc: bool, sync_type_config: dict | None = None, partitio
     return schema
 
 
-def _make_helper(*, run_maintenance_returns: int | None = None, file_uris: list[str] | None = None) -> MagicMock:
+def _make_helper(*, file_uris: list[str] | None = None) -> MagicMock:
     return MagicMock(
         get_delta_table=AsyncMock(return_value=MagicMock()),
         get_file_uris=AsyncMock(return_value=file_uris or []),
-        compact_table=AsyncMock(),
-        run_maintenance=AsyncMock(return_value=run_maintenance_returns),
     )
 
 
@@ -53,19 +52,23 @@ async def _run_post_load(
     helper: MagicMock,
     *,
     cdc_write_mode: str | None = None,
-) -> tuple[MagicMock, AsyncMock]:
+    compact_error: Exception | None = None,
+) -> tuple[AsyncMock, AsyncMock, AsyncMock]:
     job = MagicMock()
     job.id = uuid.uuid4()
     job.team_id = schema.team_id
     logger = MagicMock(adebug=AsyncMock(), ainfo=AsyncMock())
 
     prepare_s3 = AsyncMock(return_value="orders__query_1")
+    run_scheduled = AsyncMock()
+    compact_table = AsyncMock(side_effect=compact_error)
     with (
         patch(f"{_LOAD_MODULE}.prepare_s3_files_for_querying", prepare_s3),
         patch(f"{_LOAD_MODULE}.notify_revenue_analytics_that_sync_has_completed", AsyncMock()),
         patch(f"{_LOAD_MODULE}.sync_revenue_analytics_views", MagicMock()),
-        patch(f"{_LOAD_MODULE}.update_sync_type_config_keys", MagicMock()) as update_config,
         patch(f"{_LOAD_MODULE}.DataWarehouseTable", MagicMock()),
+        patch.object(DeltaMaintenance, "run_scheduled", run_scheduled),
+        patch.object(DeltaMaintenance, "compact_table", compact_table),
         patch(f"{_PIPELINE_SYNC_MODULE}.update_last_synced_at", AsyncMock()),
         patch(f"{_PIPELINE_SYNC_MODULE}.validate_schema_and_update_table", AsyncMock()),
         patch(f"{_PIPELINE_SYNC_MODULE}.register_cdc_companion_table", AsyncMock()),
@@ -82,100 +85,43 @@ async def _run_post_load(
             logger=logger,
             cdc_write_mode=cdc_write_mode,
         )
-    return update_config, prepare_s3
+    return run_scheduled, compact_table, prepare_s3
 
 
 class TestRunPostLoadDeltaMaintenance:
+    """Post-load picks the right maintenance flavor per schema kind; the threshold/watermark
+    mechanics themselves are covered in core/delta/test/test_maintenance.py."""
+
     @pytest.mark.asyncio
     async def test_cdc_schema_uses_threshold_maintenance_not_unconditional_compact(self):
         # The incident behavior this guards: CDC finals land every tick, so an unconditional
         # compact_table here means hundreds of compact+vacuum cycles per hour on a busy source.
         schema = _make_schema(is_cdc=True, sync_type_config={"last_vacuum_version": 41})
-        helper = _make_helper()
 
-        await _run_post_load(schema, helper, cdc_write_mode="incremental")
+        run_scheduled, compact_table, _ = await _run_post_load(schema, _make_helper(), cdc_write_mode="incremental")
 
-        helper.compact_table.assert_not_awaited()
-        assert helper.run_maintenance.await_args is not None
-        assert helper.run_maintenance.await_args.kwargs == {
-            "partition_count": 7,
-            "last_vacuum_version": 41,
-            "commit_threshold": 100,
-        }
-
-    @pytest.mark.asyncio
-    async def test_missing_partition_count_is_derived_from_table_layout(self):
-        # datetime/numerical-partitioned schemas persist no partition_count. Passing None through
-        # makes the threshold math treat the table as one partition, so any >200-file table would
-        # compact every tick again — the exact behavior this change removes.
-        schema = _make_schema(is_cdc=True, partition_count=None)
-        helper = _make_helper(
-            file_uris=[
-                "s3://bucket/orders/_ph_partition_key=2026-01/a.parquet",
-                "s3://bucket/orders/_ph_partition_key=2026-01/b.parquet",
-                "s3://bucket/orders/_ph_partition_key=2026-02/c.parquet",
-            ]
-        )
-
-        await _run_post_load(schema, helper, cdc_write_mode="incremental")
-
-        assert helper.run_maintenance.await_args is not None
-        assert helper.run_maintenance.await_args.kwargs["partition_count"] == 2
+        compact_table.assert_not_awaited()
+        run_scheduled.assert_awaited_once_with(schema, is_cdc_companion=False)
 
     @pytest.mark.asyncio
     async def test_non_cdc_schema_keeps_unconditional_compact(self):
         schema = _make_schema(is_cdc=False)
-        helper = _make_helper()
 
-        update_config, _ = await _run_post_load(schema, helper)
+        run_scheduled, compact_table, _ = await _run_post_load(schema, _make_helper())
 
-        helper.compact_table.assert_awaited_once()
-        helper.run_maintenance.assert_not_awaited()
-        update_config.assert_not_called()
+        compact_table.assert_awaited_once()
+        run_scheduled.assert_not_awaited()
 
     @pytest.mark.asyncio
-    async def test_cdc_companion_uses_its_own_watermark_key(self):
-        # The snapshot and _cdc companion are different delta tables with unrelated versions, so
-        # the companion must run cadence maintenance against last_vacuum_version_cdc — reading or
-        # writing the snapshot's last_vacuum_version would corrupt both cadences, and skipping
-        # cadence maintenance entirely would let companion tombstones accumulate until the
-        # file-count thresholds happen to trip. Partition count is derived from its own layout —
-        # schema.partition_count describes the snapshot table.
+    async def test_cdc_companion_write_runs_companion_maintenance(self):
+        # The snapshot and _cdc companion are different delta tables, so a companion (scd2_append)
+        # write must run maintenance in companion mode — run_scheduled then uses the companion's own
+        # watermark key and layout instead of the snapshot's (see test_maintenance.TestRunScheduled).
         schema = _make_schema(is_cdc=True, sync_type_config={"last_vacuum_version": 41, "last_vacuum_version_cdc": 7})
-        helper = _make_helper(run_maintenance_returns=9, file_uris=["s3://bucket/orders_cdc/a.parquet"])
 
-        update_config, _ = await _run_post_load(schema, helper, cdc_write_mode="scd2_append")
+        run_scheduled, _, _ = await _run_post_load(schema, _make_helper(), cdc_write_mode="scd2_append")
 
-        assert helper.run_maintenance.await_args is not None
-        assert helper.run_maintenance.await_args.kwargs == {
-            "partition_count": 1,
-            "last_vacuum_version": 7,
-            "commit_threshold": 100,
-        }
-        update_config.assert_called_once_with(schema.id, schema.team_id, updates={"last_vacuum_version_cdc": 9})
-
-    @parameterized.expand(
-        [
-            # run_maintenance returning a version must persist it — a lost watermark means
-            # vacuum_if_stale re-seeds forever and the table never vacuums.
-            ("new_version_persists", 55, True),
-            ("no_change_skips_write", None, False),
-            ("same_version_skips_write", 41, False),
-        ]
-    )
-    @pytest.mark.asyncio
-    async def test_watermark_persistence(self, _name: str, returned_version: int | None, expect_write: bool):
-        schema = _make_schema(is_cdc=True, sync_type_config={"last_vacuum_version": 41})
-        helper = _make_helper(run_maintenance_returns=returned_version)
-
-        update_config, _ = await _run_post_load(schema, helper, cdc_write_mode="incremental")
-
-        if expect_write:
-            update_config.assert_called_once_with(
-                schema.id, schema.team_id, updates={"last_vacuum_version": returned_version}
-            )
-        else:
-            update_config.assert_not_called()
+        run_scheduled.assert_awaited_once_with(schema, is_cdc_companion=True)
 
     @parameterized.expand([("non_cdc", False), ("cdc", True)])
     @pytest.mark.asyncio
@@ -187,7 +133,7 @@ class TestRunPostLoadDeltaMaintenance:
         post_maintenance_uris = ["s3://bucket/orders/compacted.parquet"]
         helper = _make_helper(file_uris=post_maintenance_uris)
 
-        _, prepare_s3 = await _run_post_load(schema, helper, cdc_write_mode="incremental" if is_cdc else None)
+        _, _, prepare_s3 = await _run_post_load(schema, helper, cdc_write_mode="incremental" if is_cdc else None)
 
         prepare_s3.assert_awaited_once()
         assert prepare_s3.await_args is not None
@@ -195,43 +141,22 @@ class TestRunPostLoadDeltaMaintenance:
 
     @parameterized.expand(
         [
-            # A genuine maintenance bug must still be captured for visibility.
-            ("genuine_bug", RuntimeError("maintenance blew up"), True),
+            # A genuine compaction bug must still be captured for visibility.
+            ("genuine_bug", RuntimeError("compaction blew up"), True),
             # A transient S3 rate-limit/connectivity blip is already non-fatal here (the next
-            # tick's maintenance retries the same idempotent cleanup) and must not be promoted
+            # sync's maintenance retries the same idempotent cleanup) and must not be promoted
             # into a fresh error-tracking issue — the regression this guards.
             ("transient_s3_slowdown", OSError("Generic S3 error: Please reduce your request rate."), False),
         ]
     )
     @pytest.mark.asyncio
-    async def test_maintenance_failure_handling(self, _name: str, error: Exception, expect_capture: bool):
-        # A maintenance hiccup must not fail the final batch — the rest of post-load
-        # (queryable folder prep, table registration) still has to run or the job wedges.
-        schema = _make_schema(is_cdc=True)
-        helper = _make_helper()
-        helper.run_maintenance = AsyncMock(side_effect=error)
-
-        with patch(f"{_LOAD_MODULE}.capture_exception") as mock_capture:
-            _, prepare_s3 = await _run_post_load(schema, helper, cdc_write_mode="incremental")
-
-        assert mock_capture.called is expect_capture
-        prepare_s3.assert_awaited_once()
-
-    @parameterized.expand(
-        [
-            ("genuine_bug", RuntimeError("compaction blew up"), True),
-            ("transient_s3_slowdown", OSError("Generic S3 error: Please reduce your request rate."), False),
-        ]
-    )
-    @pytest.mark.asyncio
     async def test_compact_failure_handling(self, _name: str, error: Exception, expect_capture: bool):
-        # Same non-fatal handling as maintenance, for the non-CDC unconditional compact_table path.
+        # A compaction hiccup must not fail the final batch — the rest of post-load
+        # (queryable folder prep, table registration) still has to run or the job wedges.
         schema = _make_schema(is_cdc=False)
-        helper = _make_helper()
-        helper.compact_table = AsyncMock(side_effect=error)
 
         with patch(f"{_LOAD_MODULE}.capture_exception") as mock_capture:
-            _, prepare_s3 = await _run_post_load(schema, helper)
+            _, _, prepare_s3 = await _run_post_load(schema, _make_helper(), compact_error=error)
 
         assert mock_capture.called is expect_capture
         prepare_s3.assert_awaited_once()
