+++ products/warehouse_sources/backend/temporal/data_imports/pipelines/core/delta/test/test_ops.py
@@ -1,0 +1,74 @@
+import pytest
+from unittest.mock import AsyncMock, MagicMock
+
+import deltalake
+
+from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.delta.ops import (
+    DELTA_MERGE_CONFLICT_RETRIES,
+    execute_with_conflict_retry,
+)
+
+
+def _make_logger():
+    return MagicMock(adebug=AsyncMock(), ainfo=AsyncMock(), awarning=AsyncMock(), aerror=AsyncMock())
+
+
+class TestExecuteWithConflictRetry:
+    """A committing operation's CommitFailedError means delta-rs's conflict checker rejected the
+    commit outright, without spending any of its own internal retry budget (see the comment on
+    DELTA_MERGE_CONFLICT_RETRIES). Regression coverage for the sync dying on the first such
+    conflict instead of refreshing the table and re-running the operation, as the error's own
+    "must be rerun" message calls for. Shared by merges and `compact_table`'s optimize.compact."""
+
+    @pytest.mark.asyncio
+    async def test_succeeds_without_retry(self):
+        table = MagicMock()
+        operation_fn = MagicMock(return_value={"num_output_rows": 1})
+
+        result = await execute_with_conflict_retry(table, operation_fn, "op", _make_logger())
+
+        assert result == {"num_output_rows": 1}
+        operation_fn.assert_called_once()
+        table.update_incremental.assert_not_called()
+
+    @pytest.mark.asyncio
+    async def test_retries_on_conflict_then_succeeds(self):
+        table = MagicMock()
+        operation_fn = MagicMock(
+            side_effect=[
+                deltalake.exceptions.CommitFailedError("Commit failed: a concurrent transactions added new data."),
+                {"num_output_rows": 1},
+            ]
+        )
+
+        result = await execute_with_conflict_retry(table, operation_fn, "op", _make_logger())
+
+        assert result == {"num_output_rows": 1}
+        assert operation_fn.call_count == 2
+        table.update_incremental.assert_called_once()
+
+    @pytest.mark.asyncio
+    async def test_gives_up_after_exhausting_retries(self):
+        table = MagicMock()
+        operation_fn = MagicMock(
+            side_effect=deltalake.exceptions.CommitFailedError(
+                "Commit failed: a concurrent transactions added new data."
+            )
+        )
+
+        with pytest.raises(deltalake.exceptions.CommitFailedError):
+            await execute_with_conflict_retry(table, operation_fn, "op", _make_logger())
+
+        assert operation_fn.call_count == DELTA_MERGE_CONFLICT_RETRIES + 1
+        assert table.update_incremental.call_count == DELTA_MERGE_CONFLICT_RETRIES
+
+    @pytest.mark.asyncio
+    async def test_other_errors_propagate_without_retry(self):
+        table = MagicMock()
+        operation_fn = MagicMock(side_effect=ValueError("not a commit conflict"))
+
+        with pytest.raises(ValueError):
+            await execute_with_conflict_retry(table, operation_fn, "op", _make_logger())
+
+        operation_fn.assert_called_once()
+        table.update_incremental.assert_not_called()
