+++ products/warehouse_sources/backend/temporal/data_imports/pipelines/core/delta_table_helper.py
@@ -5,21 +5,17 @@ from collections.abc import Callable, Sequence
 from typing import Any, Literal
 
 from django.conf import settings
-from django.db import InterfaceError, OperationalError
 
 import numpy as np
 import pyarrow as pa
 import deltalake as deltalake
 import pyarrow.compute as pc
-import posthoganalytics
-import botocore.exceptions
 import deltalake.exceptions
 from dlt.common.libs.deltalake import ensure_delta_compatible_arrow_schema
 from structlog.types import FilteringBoundLogger
 
 from posthog.exceptions_capture import capture_exception
 from posthog.sync import database_sync_to_async_pool
-from posthog.utils import get_machine_id
 
 from products.data_warehouse.backend.facade.api import aget_s3_client, ensure_bucket_exists
 from products.warehouse_sources.backend.models.external_data_job import ExternalDataJob
@@ -31,106 +27,16 @@ from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.arr
     pyarrow_schema_from_arrow_exportable,
 )
 from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.consts import PARTITION_KEY
-
-# A pre-write defensive compact fires when EITHER threshold is exceeded.
-#
-# Calibrated against the production file-count distribution (delta merge stats across
-# all teams): total files per table sit at p50≈60, p90≈470, p95≈850, with a long tail
-# (p99≈12k → ~14s merges; an observed pathological case hit ~82k files → ~45s merges).
-# Merge planning time tracks TOTAL files, not files-per-partition — delta still
-# enumerates every file's metadata even when partition pruning skips reading them — so
-# we gate on both:
-#
-# - files-per-partition: bounds per-partition fragmentation and rescues partitioned
-#   (esp. md5) tables, where a merge touches every partition. 200 sits well above the
-#   healthy steady state (compaction runs at the end of each successful sync) yet
-#   triggers long before a table reaches the slow tail.
-# - total files: a partition-count-independent backstop so a table with a high
-#   partition_count can't accumulate tens of thousands of files (each adding to merge
-#   planning time) while staying under the per-partition bar. 5,000 is above p95 (~850)
-#   — so healthy tables never trip it — and well below the p99/pathological tail.
-#
-# Tune further once the admin fragmentation view gives per-customer distributions.
-DEFAULT_COMPACT_FILES_PER_PARTITION_THRESHOLD = 200
-DEFAULT_COMPACT_TOTAL_FILES_THRESHOLD = 5000
-
-# Substrings of the object-store errors raised talking to our own S3-backed data-warehouse bucket
-# that are transient and self-recovering, not a bug in our code or a customer credential problem:
-# - the first three come from delta-rs's Rust `object_store` crate inside `DeltaTable.is_deltatable()`
-#   (IMDS/STS blips, dispatch timeouts)
-# - "Please reduce your request rate" is S3's SlowDown throttling response, surfaced by s3fs/aiobotocore
-#   when a bulk operation (e.g. `_purge_s3_prefix`'s list-then-delete) outruns the bucket's request-rate limit
-# A retry (of the same idempotent operation) clears these, so they shouldn't be treated the same as a
-# bug in our logic.
-TRANSIENT_OBJECT_STORE_ERRORS = (
-    "an error occurred while loading credentials",
-    "the credential provider was not enabled",
-    "Generic S3 error",
-    "Please reduce your request rate",
+from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.delta.errors import (
+    is_transient_object_store_error,
 )
-
-
-def is_transient_object_store_error(error: BaseException) -> bool:
-    """True for a transient object-store error, however it happened to surface.
-
-    `DeltaTable.is_deltatable()` raises these as a plain `OSError`, but table-level operations
-    (e.g. `vacuum()`, `optimize.compact()`) wrap the identical underlying object-store error text in
-    `deltalake.exceptions.DeltaError` instead — same blip, different exception type depending on
-    which delta-rs entry point hit it. `_purge_s3_prefix`'s s3fs/aiobotocore calls can also raise a
-    bare `botocore.exceptions.NoCredentialsError` unwrapped — the same IMDS/STS credential-provider
-    blip, just surfaced by aiobotocore's own credential resolution instead of delta-rs's Rust
-    `object_store` crate. `NoCredentialsError`'s message is a fixed, generic string (no needle to
-    match), but hitting our own instance-role-authenticated bucket always means the same transient
-    resolution hiccup, so it's recognized by type rather than by message.
-    """
-    if isinstance(error, botocore.exceptions.NoCredentialsError):
-        return True
-    return isinstance(error, OSError | deltalake.exceptions.DeltaError) and any(
-        needle in str(error) for needle in TRANSIENT_OBJECT_STORE_ERRORS
-    )
-
-
-# Delta's conflict checker raises CommitFailedError the moment a concurrent commit invalidates what
-# a committing operation read — a merge predicate, or optimize.compact's file-rewrite plan — unlike a
-# plain version-bump race, delta-rs does not consume max_commit_retries or retry this itself (see
-# delta-rs kernel/transaction/conflict_checker.rs), because resolving it safely requires re-reading
-# the table and re-running the operation, which is exactly what its "must be rerun" error message
-# asks the caller to do.
-DELTA_MERGE_CONFLICT_RETRIES = 3
-
-# `optimize.compact` plans its rewrite against the file list at the start of its scan, then reads
-# those files. A concurrent maintenance pass on the same table (e.g. a Temporal activity attempt that
-# heartbeat-timed-out but keeps running as a zombie — see this package's README on the equivalent
-# unfenced race for repartition) can vacuum one of those files out from under the scan before it gets
-# read, which delta-rs surfaces as this DeltaError. The scan failing here means the optimize aborted
-# before committing anything — the table is left exactly as it was, just still fragmented — so this is
-# safe to skip and retry on the next maintenance pass, not a bug in our logic.
-TRANSIENT_DELTA_MAINTENANCE_ERRORS = ("Optimize selected-file scan failed",)
-
-
-def is_transient_delta_maintenance_error(error: BaseException) -> bool:
-    return isinstance(error, deltalake.exceptions.DeltaError) and any(
-        needle in str(error) for needle in TRANSIENT_DELTA_MAINTENANCE_ERRORS
-    )
-
+from products.warehouse_sources.backend.temporal.data_imports.pipelines.core.delta.ops import (
+    execute_with_conflict_retry,
+)
 
 # _purge_s3_prefix is idempotent (every step is existence-gated), so retrying it whole after a brief
 # backoff is as safe as retrying a single failed call, and simpler.
 _PURGE_S3_PREFIX_MAX_ATTEMPTS = 4
-
-
-def is_transient_maintenance_error(error: BaseException) -> bool:
-    """Infra blips seen during pre-write maintenance that aren't a maintenance bug.
-
-    Covers S3/object-store hiccups reaching our own data-warehouse bucket (see
-    `is_transient_object_store_error` above), racy concurrent-maintenance DeltaErrors (see
-    `is_transient_delta_maintenance_error` above), and app-DB connection blips (DNS, pooler drops) hit
-    while resolving `job.folder_path()` on a pooled connection — the same `OperationalError`/`InterfaceError`
-    classification used for this failure class in `repartition_table.py`'s `_is_transient_infra_error`.
-    """
-    if isinstance(error, OperationalError | InterfaceError):
-        return True
-    return is_transient_object_store_error(error) or is_transient_delta_maintenance_error(error)
 
 
 def _delta_merge_spill_kwargs() -> dict[str, int]:
@@ -361,6 +267,18 @@ class DeltaTableHelper:
     def is_first_sync(self) -> bool:
         return self._is_first_sync
 
+    @property
+    def job(self) -> ExternalDataJob:
+        return self._job
+
+    @property
+    def resource_name(self) -> str:
+        return self._resource_name
+
+    @property
+    def logger(self) -> FilteringBoundLogger:
+        return self._logger
+
     def _get_credentials(self):
         return delta_storage_options()
 
@@ -500,28 +418,6 @@ class DeltaTableHelper:
             return []
 
         return await asyncio.to_thread(delta_table.file_uris)
-
-    async def _execute_with_conflict_retry(
-        self, table: deltalake.DeltaTable, operation_fn: Callable[[], dict], operation_name: str
-    ) -> dict:
-        """Run a Delta operation that commits (merge, optimize.compact, ...), refreshing the table
-        and re-running it on a commit conflict.
-
-        See DELTA_MERGE_CONFLICT_RETRIES for why this can't rely on delta-rs's own retry budget.
-        """
-        attempt = 0
-        while True:
-            try:
-                return await asyncio.to_thread(operation_fn)
-            except deltalake.exceptions.CommitFailedError:
-                if attempt >= DELTA_MERGE_CONFLICT_RETRIES:
-                    raise
-                attempt += 1
-                await self._logger.awarning(
-                    f"{operation_name}: commit conflict, retrying with refreshed table "
-                    f"(attempt {attempt}/{DELTA_MERGE_CONFLICT_RETRIES})"
-                )
-                await asyncio.to_thread(table.update_incremental)
 
     async def _dedupe_incremental_batch(
         self, data: pa.Table, primary_keys: Sequence[Any], use_partitioning: bool
@@ -762,8 +658,8 @@ class DeltaTableHelper:
                             .execute()
                         )
 
-                    merge_stats = await self._execute_with_conflict_retry(
-                        existing_delta_table, _do_merge, "write_to_deltalake: merge"
+                    merge_stats = await execute_with_conflict_retry(
+                        existing_delta_table, _do_merge, "write_to_deltalake: merge", self._logger
                     )
 
                     await self._logger.adebug(f"Delta Merge Stats: {json.dumps(merge_stats)}")
@@ -788,10 +684,11 @@ class DeltaTableHelper:
                         .execute()
                     )
 
-                merge_stats = await self._execute_with_conflict_retry(
+                merge_stats = await execute_with_conflict_retry(
                     existing_delta_table,
                     lambda: _do_merge_unpartitioned(data, predicate_ops),
                     "write_to_deltalake: merge",
+                    self._logger,
                 )
                 await self._logger.adebug(f"Delta Merge Stats: {json.dumps(merge_stats)}")
         elif (
@@ -945,10 +842,11 @@ class DeltaTableHelper:
                         .execute()
                     )
 
-                close_stats = await self._execute_with_conflict_retry(
+                close_stats = await execute_with_conflict_retry(
                     existing_delta_table,
                     lambda: _do_scd2_close(first_per_pk, predicate),
                     "write_scd2_to_deltalake: close merge",
+                    self._logger,
                 )
                 await self._logger.adebug(f"SCD2 close stats: {json.dumps(close_stats)}")
 
@@ -1034,173 +932,3 @@ class DeltaTableHelper:
         the metadata schema used for idempotency tagging.
         """
         return await self.has_commit_with_metadata({"run_uuid": run_uuid, "batch_index": str(batch_index)})
-
-    async def _vacuum(self, table: deltalake.DeltaTable) -> None:
-        await self._logger.adebug("Vacuuming table...")
-        vacuum_stats = await asyncio.to_thread(
-            table.vacuum, retention_hours=24, enforce_retention_duration=False, dry_run=False
-        )
-        await self._logger.adebug(json.dumps(vacuum_stats))
-
-    async def _compact(self, table: deltalake.DeltaTable) -> None:
-        await self._logger.adebug("Compacting table...")
-        compact_stats = await self._execute_with_conflict_retry(
-            table, lambda: table.optimize.compact(), "compact_table"
-        )
-        await self._logger.adebug(json.dumps(compact_stats))
-
-    async def vacuum_table(self) -> None:
-        table = await self.get_delta_table()
-        if table is None:
-            raise Exception("Deltatable not found")
-
-        await self._vacuum(table)
-
-    async def compact_table(self) -> None:
-        table = await self.get_delta_table()
-        if table is None:
-            raise Exception("Deltatable not found")
-
-        await self._compact(table)
-        # Reuse the table already resolved above instead of re-fetching it: `get_delta_table`
-        # is cached only opportunistically, so a re-fetch here can race a concurrent sync of a
-        # different table evicting this table's cache entry and spuriously report it missing.
-        await self._vacuum(table)
-        await self._logger.adebug("Compacting and vacuuming complete")
-
-    async def vacuum_if_stale(self, last_vacuum_version: int | None, commit_threshold: int) -> int | None:
-        """Vacuum tombstoned files once enough commits have accrued since the last vacuum.
-
-        Decoupled from merge success (called pre-write) so a table that OOMs its merge every run still
-        gets cleaned — the post-load compaction never runs for it, which is how tables reach ~99% dead
-        files. Vacuum only deletes dead files (an S3 LIST + delete), so unlike `compact_table`'s
-        `optimize.compact` (which rewrites partitions) it is memory-safe even on an oversized table.
-
-        Uses the delta version (commit count) as a cheap proxy for tombstone accumulation — no S3 LIST to
-        decide. Returns the current version to persist as the new watermark when it vacuumed, on first
-        encounter, or when the table was recreated (both reseed the watermark without vacuuming);
-        None when nothing changed.
-        """
-        table = await self.get_delta_table()
-        if table is None:
-            return None
-
-        version = await asyncio.to_thread(table.version)
-        if last_vacuum_version is None or version < last_vacuum_version:
-            # First encounter: seed the watermark without vacuuming so existing tables clean up gradually
-            # over the next `commit_threshold` commits rather than all vacuuming at once on deploy.
-            # A version below the watermark means the table was reset/recreated (delta versions are
-            # monotonic within one incarnation) and no reset path clears the persisted watermark —
-            # left alone it would block the cadence until the new table out-versioned the old one.
-            return version
-
-        commits_since = version - last_vacuum_version
-        if commits_since < commit_threshold:
-            await self._logger.adebug(
-                f"vacuum_if_stale: skipping, {commits_since} commits since last vacuum (< {commit_threshold})"
-            )
-            return None
-
-        await self._logger.ainfo(
-            f"vacuum_if_stale: {commits_since} commits since last vacuum (>= {commit_threshold}), vacuuming"
-        )
-        await self._vacuum(table)
-        try:
-            # Observability for the maintenance path — how often tables vacuum and how much log churn
-            # accrued between vacuums. Best-effort: telemetry must never break the sync.
-            posthoganalytics.capture(
-                distinct_id=get_machine_id(),
-                event="warehouse_delta_vacuumed",
-                properties={
-                    "team_id": self._job.team_id,
-                    "schema_id": str(self._job.schema_id),
-                    "source_id": str(self._job.pipeline_id),
-                    "resource_name": self._resource_name,
-                    "commits_since_last_vacuum": commits_since,
-                    "delta_version": version,
-                },
-            )
-        except Exception as e:
-            capture_exception(e)
-        return version
-
-    async def compact_if_fragmented(
-        self,
-        partition_count: int | None,
-        threshold: int = DEFAULT_COMPACT_FILES_PER_PARTITION_THRESHOLD,
-        total_threshold: int = DEFAULT_COMPACT_TOTAL_FILES_THRESHOLD,
-    ) -> bool:
-        """Run compact + vacuum if the table is fragmented past either threshold.
-
-        Fragmented = files-per-partition > `threshold` OR total files > `total_threshold`.
-        The total-files backstop matters because delta enumerates every file's metadata
-        during a merge even when partition pruning skips reading them, so merge planning
-        time tracks total files — a high partition_count must not let a table accumulate
-        tens of thousands of files while staying under the per-partition bar.
-
-        When `partition_count` is None it is derived from the table's actual layout (the
-        distinct file directories in the delta log, no extra I/O) — only md5 partitioning
-        persists a count on the schema, so datetime/numerical-partitioned tables always
-        arrive here with None.
-
-        Returns True if compaction ran, False if it was skipped. Cheap when the table is
-        healthy: one S3 LIST via `table.file_uris`. Intended for pre-write defensive cleanup
-        so a sync that arrived at a fragmented state (e.g. an earlier attempt that failed
-        before reaching `_post_run_operations`) cleans up before adding to the pile.
-        """
-        table = await self.get_delta_table()
-        if table is None:
-            return False
-
-        file_uris = await asyncio.to_thread(table.file_uris)
-        total_files = len(file_uris)
-        if partition_count is None:
-            # One directory per partition value; unpartitioned tables collapse to the single
-            # table root. Without this, a partitioned table with no persisted count reads as
-            # one giant partition and trips the per-partition threshold on every run.
-            partition_count = len({uri.rsplit("/", 1)[0] for uri in file_uris})
-        # Treat unpartitioned tables as one "partition" for the threshold math.
-        effective_partitions = max(partition_count or 1, 1)
-        files_per_partition = total_files / effective_partitions
-
-        fragmented = files_per_partition > threshold or total_files > total_threshold
-        stats = (
-            f"total_files={total_files}, partitions={effective_partitions}, "
-            f"files_per_partition={files_per_partition:.1f}, threshold={threshold}, "
-            f"total_threshold={total_threshold}"
-        )
-        if not fragmented:
-            await self._logger.adebug(f"compact_if_fragmented: skipping ({stats})")
-            return False
-
-        await self._logger.ainfo(f"compact_if_fragmented: triggering compact ({stats})")
-        await self._compact(table)
-        await self._vacuum(table)
-        return True
-
-    async def run_maintenance(
-        self,
-        partition_count: int | None,
-        last_vacuum_version: int | None,
-        commit_threshold: int,
-    ) -> int | None:
-        """Single pre-write maintenance entry point: compact if fragmented, else vacuum on commit cadence.
-
-        The two triggers are orthogonal — fragmentation (active file count) vs. commit cadence (tombstone
-        accrual) — but they share one outcome, the vacuum watermark. `compact_if_fragmented` already
-        vacuums as part of compaction, so when it runs it supersedes the cadence vacuum (no double vacuum
-        in one run) and the watermark advances to the post-compaction version. When nothing was fragmented,
-        fall through to `vacuum_if_stale`. Returns the single delta version to persist as the new
-        `last_vacuum_version` watermark, or None when nothing changed. Callers persist the watermark
-        via `update_sync_type_config_keys` (row-locked merge): the pre-write defensive path in
-        `common/extract.py` and the CDC post-load path in `common/load.py`.
-        """
-        compacted = await self.compact_if_fragmented(partition_count=partition_count)
-        if compacted:
-            table = await self.get_delta_table()
-            if table is None:
-                return None
-            # Compaction (which vacuumed) added a commit, advancing the version; reset the cadence
-            # watermark to it so the next vacuum is measured from this cleanup, not the old baseline.
-            return await asyncio.to_thread(table.version)
-        return await self.vacuum_if_stale(last_vacuum_version, commit_threshold)
